Windows: check whether KBFS is in use
