+++ go/install/install.go
@@ -16,7 +16,6 @@ import (
 	"github.com/kardianos/osext"
 	"github.com/keybase/client/go/libkb"
 	"github.com/keybase/client/go/logger"
-	"github.com/keybase/client/go/lsof"
 	keybase1 "github.com/keybase/client/go/protocol/keybase1"
 	"github.com/keybase/go-updater/process"
 )
@@ -258,33 +257,6 @@ func kbfsBinPathDefault(runMode libkb.RunMode, binPath string) (string, error) {
 		return "", err
 	}
 	return filepath.Join(filepath.Dir(path), kbfsBinName()), nil
-}
-
-// IsInUse returns true if the mount is in use. This may be used by the updater
-// to determine if it's safe to apply an update and restart.
-func IsInUse(mountDir string, log Log) bool {
-	log.Debug("Mount dir: %s", mountDir)
-	if mountDir == "" {
-		return false
-	}
-	if _, serr := os.Stat(mountDir); os.IsNotExist(serr) {
-		log.Debug("%s doesn't exist", mountDir)
-		return false
-	}
-
-	log.Debug("Checking mount (lsof)")
-	processes, err := lsof.MountPoint(mountDir)
-	if err != nil {
-		// If there is an error in lsof it's ok to continue
-		// An exit status of 1 means that the mount is not in use, and is
-		// not really an error.
-		// TODO: Remove this when we fix lsof
-		log.Warning("Continuing despite error in lsof: %s", err)
-	}
-	if len(processes) != 0 {
-		return true
-	}
-	return false
 }
 
 // TerminateApp will stop the Keybase (UI) app
